% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceser_main.R
\name{vcovCESE_init}
\alias{vcovCESE_init}
\title{Cluster Estimated Standard Errors}
\usage{
vcovCESE_init(x, cluster = NULL, type = c("HC", "HC1"))
}
\arguments{
\item{x}{a model object. It can be the output of the functions \code{lm} or \code{glm}.}

\item{cluster}{either a formula \code{~ rhs} with a summation expression with the name of each variable to use to cluster the data replacing the \code{rhs}, or a vector, matrix, or data.frame with the clustering variables.}

\item{type}{string with either \code{HC2} or \code{HC3} (default). Specifies the type of correction due to possible leverage effects on OLS estimates (Davidson and MacKinnon, 1993).}
}
\value{
what is returnts...
}
\description{
Compute the Cluster Estimated Standard Errors (CESE)
}
\examples{

\dontrun{
}
}
